engine/requirements.txt
fastapi
uvicorn[standard]
httpx
gitpython
anthropic
aiosqlite
python-dotenv
pydantic>=2.0
